+++ algebra/expansion-noaliasing.h
@@ -403,7 +403,7 @@ OutputStream& operator<<(OutputStream& os, const typename Expansion<X>::referenc
     return os << "<reference>";
 }
 
-OutputStream& operator<<(OutputStream& os, const Expansion<Float>::const_reference& cref) {
+OutputStream& operator<<(OutputStream& os, const Expansion<Float64>::const_reference& cref) {
     //return os << "(" << v.get<0>() << ":" << v.get<1>() << ")";
     return os << "<const_reference>";
 }
@@ -528,8 +528,8 @@ Expansion<X> embed(unsigned int before_size, const Expansion<X>& x, unsigned int
 }
 
 
-inline Expansion<Float> midpoint(const Expansion<ExactInterval>& pse) {
-    Expansion<Float> r(pse.argument_size());
+inline Expansion<Float64> midpoint(const Expansion<ExactInterval>& pse) {
+    Expansion<Float64> r(pse.argument_size());
     for(Expansion<ExactInterval>::ConstIterator iter=pse.begin(); iter!=pse.end(); ++iter) {
         //r.append(iter->key(),midpoint(iter->data())); }
         r.append(iter->key(),midpoint(iter->data())); }
@@ -593,7 +593,7 @@ Vector<Y> evaluate(const Vector< Expansion<X> >& x, const Vector<Y>& y)
 }
 
 
-template<class X> Vector< Expansion<X> > operator*(const Expansion<X>& e, const Vector<Float> v) {
+template<class X> Vector< Expansion<X> > operator*(const Expansion<X>& e, const Vector<Float64> v) {
     Vector< Expansion<X> > r(v.size(),Expansion<X>(e.argument_size()));
     for(Nat i=0; i!=r.size(); ++i) {
         ARIADNE_ASSERT(v[i]==0.0 || v[i]==1.0);
@@ -603,8 +603,8 @@ template<class X> Vector< Expansion<X> > operator*(const Expansion<X>& e, const 
 }
 
 
-inline Vector< Expansion<Float> > midpoint(const Vector< Expansion<ExactInterval> >& pse) {
-    Vector< Expansion<Float> > r(pse.size());
+inline Vector< Expansion<Float64> > midpoint(const Vector< Expansion<ExactInterval> >& pse) {
+    Vector< Expansion<Float64> > r(pse.size());
     for(Nat i=0; i!=pse.size(); ++i) {
         r[i]=midpoint(pse[i]); }
     return r;
