+++ algebra/multi_index-noaliasing.h
@@ -47,7 +47,7 @@ uint32_t fac(uint8_t);
 uint32_t bin(uint8_t,uint8_t);
 
 typedef unsigned char uchar;
-typedef double Float;
+typedef double Float64;
 
 class MultiIndexReference;
 class MultiIndexValueReference;
